+++ src/main/java/HawkerUncle/command/AddCommand.java
@@ -1,3 +1,14 @@
+package HawkerUncle.command;
+
+import HawkerUncle.storage.Storage;
+import HawkerUncle.ui.Ui;
+import HawkerUncle.task.TaskList;
+import HawkerUncle.task.Task;
+import HawkerUncle.task.ToDo;
+import HawkerUncle.task.Deadline;
+import HawkerUncle.task.Event;
+import HawkerUncle.parser.Parser;
+
 import java.util.ArrayList;
 import java.time.LocalDateTime;
