+++ src/main/java/HawkerUncle/command/ByeCommand.java
@@ -1,4 +1,10 @@
-public class ByeCommand implements Command{
+package HawkerUncle.command;
+
+import HawkerUncle.storage.Storage;
+import HawkerUncle.task.TaskList;
+import HawkerUncle.ui.Ui;
+
+public class ByeCommand implements Command {
     @Override
     public void execute(TaskList task, Ui ui, Storage storage){
         System.out.println("Bye. Hope to see you again soon!");
