+++ src/main/java/HawkerUncle/command/Command.java
@@ -1,3 +1,9 @@
+package HawkerUncle.command;
+
+import HawkerUncle.storage.Storage;
+import HawkerUncle.task.TaskList;
+import HawkerUncle.ui.Ui;
+
 public interface Command {
     public abstract void execute(TaskList tasks, Ui ui, Storage storage);
     public abstract boolean isExit();
