+++ src/main/java/HawkerUncle/command/ListCommand.java
@@ -1,4 +1,10 @@
-public class ListCommand implements Command{
+package HawkerUncle.command;
+
+import HawkerUncle.storage.Storage;
+import HawkerUncle.task.TaskList;
+import HawkerUncle.ui.Ui;
+
+public class ListCommand implements Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         System.out.println("  Here are the tasks in your list:");
