+++ src/main/java/HawkerUncle/command/MarkCommand.java
@@ -1,15 +1,20 @@
+package HawkerUncle.command;
+
+import HawkerUncle.storage.Storage;
+import HawkerUncle.task.TaskList;
+import HawkerUncle.ui.Ui;
+
 import java.io.IOException;
 
-public class DeleteCommand implements Command{
+public class MarkCommand implements Command {
     private final int idx;
-    public DeleteCommand(int idx) {
+    public MarkCommand(int idx) {
         this.idx = idx;
     }
-
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         if (idx < 0 || idx >= tasks.size()) throw new IndexOutOfBoundsException("Task number is invalid.");
-        Task removed = tasks.remove(idx);
+        tasks.get(idx).setDone(true);
 
         try {
             storage.save(tasks);
@@ -17,11 +22,9 @@ public class DeleteCommand implements Command{
             System.out.println("Task is not saved");
         }
 
-        System.out.println("  Noted. I've removed this task:");
-        System.out.println("    " + removed);
-        System.out.println("  Now you have " + tasks.size() + " tasks in the list");
+        System.out.println("  Nice! I've marked this task as done:");
+        System.out.println("  " + tasks.get(idx));
     }
-
     @Override
     public boolean isExit() {
         return false;
