+++ src/main/java/HawkerUncle/storage/Storage.java
@@ -1,3 +1,10 @@
+package HawkerUncle.storage;
+import HawkerUncle.task.TaskList;
+import HawkerUncle.task.Task;
+import HawkerUncle.task.ToDo;
+import HawkerUncle.task.Deadline;
+import HawkerUncle.task.Event;
+
 import java.io.File;
 import java.io.IOException;
 import java.io.FileWriter;
