+++ src/main/java/HawkerUncle/task/Task.java
@@ -1,3 +1,5 @@
+package HawkerUncle.task;
+
 public abstract class Task { //partially taken from Level 3 A-Classes at https://nus-cs2103-ay2526s1.github.io/website/schedule/week2/project.html
     protected String description;
     protected boolean isDone;
@@ -16,5 +18,8 @@ public abstract class Task { //partially taken from Level 3 A-Classes at https:/
         return "[" + getStatusIcon() + "] " + this.description;
     }
 
+    public void setDone(boolean isDone) {
+        this.isDone = isDone;
+    }
     public abstract String toSaveFormat();
 }
